requirements.txt
nltk
scikit-learn
gensim
numpy
tensorflow-cpu
kagglehub
